dashboard/backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
websockets==11.0.1
pydantic==1.10.7
python-socketio[client]==5.8.0
fastapi-socketio==0.0.10
aiofiles==23.1.0
psutil==5.9.5
pandas==2.0.0
numpy==1.24.3
scikit-learn==1.2.2
python-dateutil==2.8.2
pyzmq==25.0.2
